+++ src/boomerang/db/exp.cpp
@@ -2311,7 +2311,7 @@ void FlagDef::appendDotFile(QTextStream& of)
 	of << "opFlagDef \\n" << HostAddress(this) << "| ";
 	// Display the RTL as "RTL <r1> <r2>..." vertically (curly brackets)
 	of << "{ RTL ";
-	int n = rtl->size();
+	size_t n = rtl->size();
 
 	for (int i = 0; i < n; i++) {
 		of << "| <r" << i << "> ";
